Add preliminary framework for paramlist support.
Add the beginnings of a system to process and pass parameterlists
to the Ri calls. Only very early, only supports the additional
arguments approach, where each parameter consists of a name and
value, which can be a single value of the right type, or a list()
of values. Currently only coded support for non-array floats.

Add code to use Aqsis' CqTokenDictionary for parsing and recording
token declarations.