Change API naming convention, and initialisation code.
Update the Ri API naming convention to better match Io. Instead
of RiWorldBegin, it is now worldBegin.

Change the way RenderMan object initialisation works, now defautl
output is RI_NULL (console) and a new function 'with' allows
cloning with a different output name to RiBegin. Not finished yet
as it seems the dangling contexts mean it outputs extra RIB
streams to the console on exit.