Update parameter type handling.
Add handling of integer and string parameter types, found
when adding extra code to the test1.io test case.

Remove mistakenly left in debug output message.

Change the range of API calls that are converted by the xsl
transform, used to ignore things with CustomImpl, but that
resulted in too many, as the flag was really meant for RIB
generation in Aqsis. At the moment, there are probably API
calls that are being incorrectly generated, when a proper test
setup is completed, these can be identified and updated.