Implement basic RiProcedural support.
RiProcedural is now implemented in the same style as RIB. A string
identifies the procedural type from the three built-in types. A
string array defines the arguments to the procedural, which is then
built into a form suitable for passing over the standard
RiProcedural API function, and for cleaning up using the standard
RiProcFree.